삭제 확인창 + 삭제 알림 메시지 추가

+++ src/components/Carlist.js
@@ -1,14 +1,29 @@
 import React, {useEffect, useState} from "react";
 import {SERVER_URL} from "../constants";
 import {DataGrid} from "@mui/x-data-grid";
+import Snackbar from "@mui/material/Snackbar";
 
 function Carlist() {
   const [cars, setCars] = useState([]);
 
   const onDelClick = (url) => {
-    fetch(url, {method: 'DELETE'})
-            .then(response => fetchCars())
+
+    if (window.confirm("Are you sure to delete?")) {
+      fetch(url, {method: 'DELETE'})
+              .then(response => {
+                if (response.ok) {
+                  fetchCars();
+                  setOpen(true);
+                } else {
+                  alert('Something went wrong!')
+                }
+              })
+              .catch(err => console.error(err))
+    }
   }
+
+  const [open, setOpen] = useState(false);
+
   const columns = [
     {field: 'brand', headerName: 'Brand', width: 200},
     {field: 'model', headerName: 'Model', width: 200},
@@ -42,7 +57,13 @@ function Carlist() {
             <DataGrid
                     rows={cars}
                     columns={columns}
+                    disableSelectionOnClick={true}
                     getRowId={row => row._links.self.href}/>
+            <Snackbar
+                    open={open}
+                    autoHideDuration={2000}
+                    onClose={() => setOpen(false)}
+                    message={"Car deleted"}/>
           </div>
   )
 }
